requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==4.4.2
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.1
httptools==0.5.0
httpx==0.24.1
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.24.3
Pillow==9.5.0
proglog==0.1.10
pydantic==1.10.7
python-dotenv==1.0.0
python-telegram-bot==20.3
pytube==15.0.0
PyYAML==6.0
requests==2.30.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
